Created stub for class representing Wofol city scroll and tested it
A scroll that transports the player to Wofol city, WofolCityScroll. This
extends the Marked scroll class which is a class that
can be marked with any destination in the stendhal game. Also added
method in it that allows calling the protected "useScroll" function
in the superclass of MarkedScroll i.e TeleportScroll.

Wrote two failing acceptance tests, one to check the coordinates
and zone of the play change correctly and another to check the
return value of the describe function in WofolCityScroll.

Issue: A Teleportation scroll for Wofol City